# pedidos-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de pedidos",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/products/{id_product}/upload-file": {
            "post": {
                "summary": "Upload files for a product",
                "description": "Uploads files for a specific product.",
                "operationId": "a82adb3cc56d09bf2431309c825dcef7",
                "parameters": [
                    {
                        "name": "id_product",
                        "in": "path",
                        "description": "ID of the product",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Form data containing files",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "files"
                                ],
                                "properties": {
                                    "files": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    },
                                    "id_user": {
                                        "description": "ID of the user",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Files uploaded successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Archivos guardados correctamente"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error or user/product not found",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "message": "Validation error details or User/Producto no encontrada"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}